+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 1.0.3 *(2020-08-20)*
+----------------------------
+
+ * Upgrade Kotlin, Gradle, plugins, dependencies
+ * Move project modules config into common file
+
 Version 1.0.2 *(2020-08-16)*
 ----------------------------